+++ include/platform/hal_semaphore.h
@@ -14,6 +14,8 @@
 
 #include "ud3tn/result.h"
 
+#include <stdint.h>
+
 /**
  * @brief hal_semaphore_init_binary Creates a new binary semaphore meaning that
  *				  only one execution entity can hold at a
@@ -62,11 +64,13 @@ void hal_semaphore_delete(Semaphore_t sem);
  * @brief hal_semaphore_try_take Try to take a initialized semaphore, abort
  *				 after the timeout is reached and return
  *				 an error
- * @param sem The semaphore that should be locket
- * @param timeout_ms The timeout in Milliseconds
- * @return Whether the operation was successful
+ * @param sem The semaphore that should be locked
+ * @param timeout_ms The timeout in Milliseconds, in the range of 0 to
+ *		     9223372036854. If outside of this range, the function is
+ *		     equivalent to hal_semaphore_take_blocking
+ * @return Whether the operation was successful; UD3TN_FAIL on timeout
  */
-enum ud3tn_result hal_semaphore_try_take(Semaphore_t sem, int timeout_ms);
+enum ud3tn_result hal_semaphore_try_take(Semaphore_t sem, int64_t timeout_ms);
 
 
 #endif /* HAL_SEMAPHORE_H_INCLUDED */
